Extracted PathNode to separate file
