+++ lib/arel/transformer.rb
@@ -1,5 +1,6 @@
 require_relative './transformer/node'
 require_relative './transformer/path'
+require_relative './transformer/path_node'
 require_relative './transformer/visitor'
 
 module Arel
